Refactor predicate test files to own folder
