+++ src/test/java/staffconnect/model/person/predicates/PersonHasModulePredicateTest.java
@@ -1,4 +1,4 @@
-package staffconnect.model.person;
+package staffconnect.model.person.predicates;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -6,7 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import staffconnect.model.person.predicates.PersonHasModulePredicate;
+import staffconnect.model.person.Module;
 import staffconnect.testutil.PersonBuilder;
 
 public class PersonHasModulePredicateTest {
